init pawtograder-exec; setup nix build
pyret-autograder-pawtograder-exec is the machine facing cli that is
meant to be used by the pawtograder platform. Unlike the prior solution,
the action doesn't need to have any special environment except having
the `pyret-pawtograder` executable in their $PATH. This will allow
discontinuing our fork of the action and upstream the minimal changes
needed to support this cli.

The long term goal being that this repo can produce docker images using
nix that can be used by the action.

Eventually, pawtograder-exec should also be used by the more human
friendly pyret-autograder-cli.